TEIID-4493 initial qgis support # Conflicts: #	connectors/translator-jdbc/src/main/java/org/teiid/translator/jdbc/teiid/TeiidExecutionFactory.java #	engine/src/main/java/org/teiid/query/QueryPlugin.java #	engine/src/main/java/org/teiid/query/resolver/util/ResolverVisitor.java #	engine/src/main/java/org/teiid/query/rewriter/QueryRewriter.java #	engine/src/main/java/org/teiid/query/validator/ValidationVisitor.java #	engine/src/test/java/org/teiid/query/processor/TestGeometry.java #	runtime/src/main/java/org/teiid/deployers/PgCatalogMetadataStore.java #	runtime/src/main/java/org/teiid/odbc/ODBCServerRemoteImpl.java #	runtime/src/main/java/org/teiid/transport/PgBackendProtocol.java #	runtime/src/main/java/org/teiid/transport/PgFrontendProtocol.java #	test-integration/common/src/test/resources/TestCase3473/testGetTables.expected #	test-integration/common/src/test/resources/TestJDBCSocketTransport/testSelect.expected #	test-integration/common/src/test/resources/TestMMDatabaseMetaData/testGetColumns.expected #	test-integration/common/src/test/resources/TestMMDatabaseMetaData/testGetFunctionColumns.expected #	test-integration/common/src/test/resources/TestMMDatabaseMetaData/testGetFunctions.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_ATTRDEF.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_ATTRIBUTE.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_ATTRIBUTE_overflow.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_CLASS.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_INDEX.expected #	test-integration/common/src/test/resources/TestODBCSchema/test_PG_NAMESPACE.expected #	test-integration/common/src/test/resources/TestODBCSocketTransport/testSelect.expected #	test-integration/common/src/test/resources/TestPartsDatabaseMetadata/testColumns.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testColumns.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testFunctionParams.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testFunctions.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testKeyColumns.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testKeys.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testProperties.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testTables.expected #	test-integration/common/src/test/resources/TestSystemVirtualModel/testViews.expected #	test-integration/common/src/test/resources/TestVirtualDocWithVirtualProc/testDefect15241b.expected

+++ runtime/src/main/java/org/teiid/odbc/ODBCServerRemote.java
@@ -34,7 +34,7 @@ public interface ODBCServerRemote {
 	
 	void prepare(String prepareName, String sql, int[] paramType);
 
-	void bindParameters(String bindName, String prepareName, Object[] paramdata, int resultCodeCount, int[] resultColumnFormat);
+	void bindParameters(String bindName, String prepareName, Object[] paramdata, int resultCodeCount, short[] resultColumnFormat);
 	
 	void execute(String bindName, int maxrows);
